separate method injector
